+++ compressutil/compressutil.go
@@ -1,4 +1,5 @@
-package compress
+// Package compressutil a plugin to handle compression and decompression tasks
+package compressutil
 
 import (
 	"bytes"
@@ -7,12 +8,14 @@ import (
 	"io"
 )
 
+// ReaderError for notating that an error occurred while reading compressed data.
 type ReaderError struct {
 	error
 
 	cause error
 }
 
+// Decode data in a bytes.Reader based on supplied encoding.
 func Decode(byteReader *bytes.Reader, encoding string) (data []byte, err error) {
 	reader, err := getRawReader(byteReader, encoding)
 	if err != nil {
@@ -35,8 +38,10 @@ func getRawReader(byteReader *bytes.Reader, encoding string) (io.Reader, error) 
 	}
 }
 
-func Encode(data []byte, encoding string) (readCloser io.ReadCloser, err error, encoded bool) {
+// Encode data in a []byte based on supplied encoding.
+func Encode(data []byte, encoding string) (readCloser io.ReadCloser, err error) {
 	byteBuffer := new(bytes.Buffer)
+
 	compressor, err := getCompressor(byteBuffer, encoding)
 	if err != nil {
 		// If an error creating the compressor occurs set to nil to allow uncompressed data to be returned
@@ -44,14 +49,14 @@ func Encode(data []byte, encoding string) (readCloser io.ReadCloser, err error, 
 	}
 
 	if compressor != nil {
-		compressor.Write(data)
+		_, _ = compressor.Write(data)
 
 		if err = compressor.Close(); err == nil {
-			return io.NopCloser(byteBuffer), nil, true
+			return io.NopCloser(byteBuffer), nil
 		}
 	}
 
-	return io.NopCloser(bytes.NewReader(data)), err, false
+	return nil, err
 }
 
 func getCompressor(byteBuffer *bytes.Buffer, encoding string) (compressor io.WriteCloser, err error) {
